+++ src/app/app.module.ts
@@ -3,6 +3,7 @@ import {NgModule} from '@angular/core';
 import {MatButtonModule} from '@angular/material/button';
 import {MatNativeDateModule} from '@angular/material/core';
 import {MatDatepickerModule} from '@angular/material/datepicker';
+import {MatDialogModule} from '@angular/material/dialog';
 import {
   MAT_FORM_FIELD_DEFAULT_OPTIONS,
   MatFormFieldDefaultOptions,
@@ -35,6 +36,7 @@ import {
     // Angular Material
     MatButtonModule,
     MatDatepickerModule,
+    MatDialogModule,
     MatIconModule,
     MatListModule,
     MatNativeDateModule,