+++ src/app/core/services/models/last-commit-info.ts
@@ -1,0 +1,6 @@
+export interface LastCommitInfo {
+  envLastCommitInfo: string | null;
+  envBuildDate: string | null;
+  jarLastCommitInfo: string | null;
+  jarBuildDate: string | null;
+}