+++ src/app/routes/connector-ui/dashboard-page/dashboard-page/dashboard-page-data.ts
@@ -1,3 +1,4 @@
+import {PropertyGridGroup} from '../../../../component-library/property-grid/property-grid-group/property-grid-group';
 import {Fetched} from '../../../../core/services/models/fetched';
 import {DonutChartData} from '../dashboard-donut-chart/donut-chart-data';
 
@@ -10,6 +11,7 @@ export interface DashboardPageData {
   numContractDefinitions: Fetched<number>;
   numPolicies: Fetched<number>;
   numCatalogs: Fetched<number>;
+  connectorProperties: PropertyGridGroup[];
 }
 
 export function defaultDashboardData(): DashboardPageData {
@@ -22,5 +24,6 @@ export function defaultDashboardData(): DashboardPageData {
     numPolicies: Fetched.empty(),
     numContractDefinitions: Fetched.empty(),
     numCatalogs: Fetched.empty(),
+    connectorProperties: [],
   };
 }